updating activity feed screen structure
